requirements.txt
python-telegram-bot>=21.6
psycopg[binary,pool]>=3.2
python-dotenv>=1.0
